+++ functional-definitions/functional-compiler/src/main/java/com/dan323/functional/annotation/compiler/internal/MonoidCompiler.java
@@ -57,7 +57,7 @@ public final class MonoidCompiler implements Compiler {
             }
         }
         if (!success) {
-            error("The static public function op or unit was not found in %s", element.getQualifiedName());
+            error("The public function op or unit was not found in %s", element.getQualifiedName());
         }
     }